requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
APScheduler==3.11.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.8.0
cryptography==46.0.3
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
msgspec==0.19.0
multidict==6.7.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
playwright==1.56.0
propcache==0.4.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2-binary==2.9.11
pycparser==2.23
pyee==13.0.0
PyJWT==2.10.1
pypdfium2==5.1.0
python-barcode==0.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
reportlab==4.0.7
requests==2.31.0
six==1.17.0
soupsieve==2.8
twilio==8.10.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.0.1
xlrd==2.0.2
yarl==1.22.0
